@@ -1,15 +1,19 @@
-import React, { Component } from 'react';
-import { hot } from 'react-hot-loader';
-import './App.css';
+// routes
+import Router from './routes';
+// theme
+import ThemeProvider from './theme';
+// components
+import ScrollToTop from './components/scroll-to-top';
+import { StyledChart } from './components/chart';
 
-class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <h1> Hello, World!!! </h1>
-      </div>
-    );
-  }
+// ----------------------------------------------------------------------
+
+export default function App() {
+  return (
+    <ThemeProvider>
+      <ScrollToTop />
+      <StyledChart />
+      <Router />
+    </ThemeProvider>
+  );
 }
-
-export default hot(module)(App);
